friendship requests
users are able to send requests, accept requests, and unfriend other users.
there also is a new "users" page that shows all the active users.
at the moment, the functions are all working and affect the database, but there is an issue where if you click on the button, the page crashes.